pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "komodora", email = "724komodora915@gmail.com" }]
dependencies = [
    "django-debug-toolbar==4.2.0",
    "djangorestframework==3.14.0",
    "drf-spectacular==0.26.5",
    "drf-writable-nested==0.7.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "NONE" }

[tool.pdm.dev-dependencies]
dev = ["ruff>=0.4.2"]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
django = "python src/sandbox/manage.py"
start = "python src/sandbox/manage.py runserver"
lint = "ruff check"
format = "ruff format"
test = "python src/sandbox/manage.py test src/tests"

[tool.ruff]
exclude = ["src/sandbox/usage/migrations"]

[tool.ruff.lint]
select = ["F", "E", "W", "C90", "I", "N", "UP", "S", "DJ"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["usage"]

[tool.ruff.lint.per-file-ignores]
"src/tests/**" = ["N802"]
